requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
etherscan-python==2.1.0
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-cache==0.6.0
six==1.15.0
url-normalize==1.4.3
urllib3==1.26.4
waitress==2.0.0
Werkzeug==1.0.1
